Criação parcial do banco de dados

+++ CentralTicket/Contexts/Billing/ValueObjects/Price.cs
@@ -2,9 +2,9 @@
 {
     public class Price
     {
-        public float Value { get; private set; }
+        public decimal Value { get; private set; }
 
-        public Price(float price)
+        public Price(decimal price)
         {
             if (price < 0) throw new Exception("Valor deve ser positivo");
